// src/app/models/dummy.ts
export class DummyData{
    "id": number | undefined;
    "employee_name": string | undefined;
    "employee_salary": number | undefined;
    "employee_age": number | undefined;
    "profile_image": string | undefined;
}

export class DummyUpdateData{
    "id": number | undefined;
    "name": string | undefined;
    "salary": string | undefined;
    "age": string | undefined;
}